Ensure SMS and MMS messages are grouped into the same conversation.
For MMS messages, this requires knowing which of the message participants is the owner of the archive,
so that we can exclude their address from grouping (since their address is never included for SMS messages).
The address appears to be excluded from the "address" attribute and only included in the "addr" elements,
so we can use that.